some quick update for core service

+++ core-service/src/main/java/com/structura/steel/coreservice/service/impl/SaleOrderDetailServiceImpl.java
@@ -1,10 +1,13 @@
 package com.structura.steel.coreservice.service.impl;
 
+import com.structura.steel.commons.exception.ResourceNotBelongToException;
 import com.structura.steel.commons.exception.ResourceNotFoundException;
 import com.structura.steel.commons.response.PagingResponse;
+import com.structura.steel.coreservice.entity.SaleOrder;
 import com.structura.steel.coreservice.entity.SaleOrderDetail;
 import com.structura.steel.coreservice.mapper.SaleOrderDetailMapper;
 import com.structura.steel.coreservice.repository.SaleOrderDetailRepository;
+import com.structura.steel.coreservice.repository.SaleOrderRepository;
 import com.structura.steel.coreservice.service.SaleOrderDetailService;
 import com.structura.steel.dto.request.SaleOrderDetailRequestDto;
 import com.structura.steel.dto.response.SaleOrderDetailResponseDto;
@@ -22,11 +25,16 @@ import java.util.stream.Collectors;
 public class SaleOrderDetailServiceImpl implements SaleOrderDetailService {
 
     private final SaleOrderDetailRepository saleOrderDetailRepository;
+    private final SaleOrderRepository saleOrderRepository;
+
     private final SaleOrderDetailMapper saleOrderDetailMapper;
 
     @Override
     public SaleOrderDetailResponseDto createSaleOrderDetail(SaleOrderDetailRequestDto dto, Long saleId) {
+        SaleOrder saleOrder = saleOrderRepository.findById(saleId).orElseThrow(() -> new ResourceNotFoundException("SaleOrder", "id", saleId));
         SaleOrderDetail saleOrderDetail = saleOrderDetailMapper.toSaleOrderDetail(dto);
+        saleOrderDetail.setSaleOrder(saleOrder);
+
         SaleOrderDetail savedSaleOrderDetail = saleOrderDetailRepository.save(saleOrderDetail);
         return saleOrderDetailMapper.toSaleOrderDetailResponseDto(savedSaleOrderDetail);
     }
@@ -35,6 +43,12 @@ public class SaleOrderDetailServiceImpl implements SaleOrderDetailService {
     public SaleOrderDetailResponseDto updateSaleOrderDetail(Long id, SaleOrderDetailRequestDto dto, Long saleId) {
         SaleOrderDetail existing = saleOrderDetailRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("SaleOrderDetail", "id", id));
+
+        SaleOrder saleOrder = saleOrderRepository.findById(saleId).orElseThrow(() -> new ResourceNotFoundException("SaleOrder", "id", saleId));
+        if(!saleOrder.getSaleOrderDetails().contains(existing)) {
+            throw new ResourceNotBelongToException("Sale detail", "id", id, "sale order", "id", saleId);
+        }
+
         saleOrderDetailMapper.updateSaleOrderDetailFromDto(dto, existing);
         SaleOrderDetail updated = saleOrderDetailRepository.save(existing);
         return saleOrderDetailMapper.toSaleOrderDetailResponseDto(updated);
@@ -44,6 +58,11 @@ public class SaleOrderDetailServiceImpl implements SaleOrderDetailService {
     public SaleOrderDetailResponseDto getSaleOrderDetailById(Long id, Long saleId) {
         SaleOrderDetail saleOrderDetail = saleOrderDetailRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("SaleOrderDetail", "id", id));
+
+        SaleOrder saleOrder = saleOrderRepository.findById(saleId).orElseThrow(() -> new ResourceNotFoundException("SaleOrder", "id", saleId));
+        if(!saleOrder.getSaleOrderDetails().contains(saleOrderDetail)) {
+            throw new ResourceNotBelongToException("Sale detail", "id", id, "sale order", "id", saleId);
+        }
         return saleOrderDetailMapper.toSaleOrderDetailResponseDto(saleOrderDetail);
     }
 
@@ -51,6 +70,11 @@ public class SaleOrderDetailServiceImpl implements SaleOrderDetailService {
     public void deleteSaleOrderDetailById(Long id, Long saleId) {
         SaleOrderDetail saleOrderDetail = saleOrderDetailRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("SaleOrderDetail", "id", id));
+
+        SaleOrder saleOrder = saleOrderRepository.findById(saleId).orElseThrow(() -> new ResourceNotFoundException("SaleOrder", "id", saleId));
+        if(!saleOrder.getSaleOrderDetails().contains(saleOrderDetail)) {
+            throw new ResourceNotBelongToException("Sale detail", "id", id, "sale order", "id", saleId);
+        }
         saleOrderDetailRepository.delete(saleOrderDetail);
     }
